mps initialization split files and clear docs
