+++ docs/source/mps/init.rst
@@ -40,9 +40,8 @@ An empty MPS/MPO can be filled with tensors by setting them one by one.
 
 Tensor should be of the rank expected for :ref:`MPS<theory/mps/basics:Matrix product state (MPS)>` or :ref:`MPO<theory/mps/basics:Matrix product operator (MPO)>`.
 The virtual dimensions/spaces of the neighboring MPS/MPO tensors should be consistent, which, however, is not tested during direct assigment.
-The examples of creating MPS/MPO by hand can be found here:
-:ref:`Ground state of Spin-1 AKLT model<examples/mps/build:Ground state of spin-1 AKLT model>`, and
-:ref:`MPO for hopping model with U(1) symmetry<examples/mps/build:Hamiltonian for nearest-neighbor hopping/XX model>`.
+For examples creating MPS/MPO by hand, see :ref:`Ground state of Spin-1 AKLT model<examples/mps/build:Ground state of spin-1 AKLT model>`
+and :ref:`MPO for hopping model with U(1) symmetry<examples/mps/build:Hamiltonian for nearest-neighbor hopping/XX model>`.
 
 
 Initializing product MPS/MPO
@@ -61,56 +60,40 @@ Initializing random MPS/MPO
 .. autofunction:: yastn.tn.mps.random_mpo
 
 
-Making a copy of MPS/MPO
-------------------------
-
-To create an independent copy or clone of MPS/MPO :code:`psi` call :code:`psi.copy()`
-or :code:`psi.clone()`, respectively.
-It is also possible to make a shallow copy with :code:`psi.shallow_copy()`, where explicit copies of data tensors are not created.
-
-.. autofunction:: yastn.tn.mps.MpsMpo.shallow_copy
-
-.. autofunction:: yastn.tn.mps.MpsMpo.copy
-
-.. autofunction:: yastn.tn.mps.MpsMpo.clone
-
-
 Generating MPO using Hterm
 --------------------------
 
-:class:`yastn.tn.mps.Hterm` is a basic building block to define operators.
-Each ``Hterm`` represents a product of local (on-site) operators.
+We provide functionality to build MPO representations for a broad class of
+operators, e.g., Hamiltonians, given as a sum of products of local (on-site) operators.
+They are encoded as a list of :class:`mps.Hterm<yastn.tn.mps.Hterm>`, where each :class:`mps.Hterm<yastn.tn.mps.Hterm>`
+represents a product of local operators, including the numerical coefficient (amplitude) in front of that product.
 
 .. autoclass:: yastn.tn.mps.Hterm
 
-.. note::
-     :code:`Hterm.operators` should be a list of matrices with signatutes :math:`s=(1, -1)`.
-
-A list(Hterm) defines a broad class of operators of interests.
-In order to generate the corresponding MPO use :code:`mps.generate_mpo(I, terms)`,
-where :code:`terms` is a list of Hterm-s and :code:`I` is the identity MPO.
-The latter can conveniently be created with a :meth:`generator<yastn.tn.mps.Generator.I>` described below.
-An example using this method can be found :ref:`here<examples/mps/build:Building MPO using Hterm>`.
+To generate the corresponding MPO use :code:`mps.generate_mpo(I, terms)`,
+where `terms` is a list of Hterm-s and `I` is the identity MPO.
+The latter can be conveniently created as shown in :class:`mps.product_mpo<yastn.tn.mps.product_mpo>`.
+For an example, see :ref:`MPO for spinless fermions with hopping at arbitrary range using Hterms<examples/mps/build:Building MPO using Hterm>`.
 
 .. autofunction:: yastn.tn.mps.generate_mpo
 
 .. autofunction:: yastn.tn.mps.generate_mpo_fast
 
-.. autofunction:: yastn.tn.mps.generate_mpo_template
+.. autofunction:: yastn.tn.mps.generate_mpo_preprocessing
 
 
-Generator class for MPO/MPS
----------------------------
+Generator class for MPO/MPS (beta)
+----------------------------------
 
-:class:`yastn.tn.mps.Generator` automatizes the creation of MPS and MPO.
-Given a set of local (on-site) operators, e.g. :class:`yastn.operators.Spin12`
-one can build both the states and operators. The MPS/MPO can be given as a LaTeX-like expression.
+A class supporting automatizes generation of MPOs from LaTeX-like expressions.
 
 .. autoclass:: yastn.tn.mps.Generator
 
 We can directly output identity MPO built from the identity `I` from the operator generator class.
 
 .. autofunction:: yastn.tn.mps.Generator.I
+
+.. autofunction:: yastn.tn.mps.Generator.mpo_from_latex
 
 Generator supports latex-like string instructions to help building MPOs.
 For examples, see :ref:`Generate MPO from LaTex<examples/mps/build:Generator class for MPO/MPS>`.
@@ -126,11 +109,26 @@ It also provides a direct link to a random number generator in the backend to fi
 .. autofunction:: yastn.tn.mps.Generator.random_seed
 
 
+Making a copy of MPS/MPO
+------------------------
+
+To create an independent copy or clone of MPS/MPO :code:`psi` call :code:`psi.copy()`
+or :code:`psi.clone()`, respectively.
+It is also possible to make a shallow copy with :code:`psi.shallow_copy()`, where explicit copies of data tensors are not created.
+
+.. autofunction:: yastn.tn.mps.MpsMpo.shallow_copy
+
+.. autofunction:: yastn.tn.mps.MpsMpo.copy
+
+.. autofunction:: yastn.tn.mps.MpsMpo.clone
+
+
 Import and export MPS/MPO from/to different formats
 ---------------------------------------------------
 
-MPS/MPO can save as Python `dict` or `HDF5`` file.
-The MPS/MPO previously serialized by :meth:`yastn.tn.mps.MpsMpo.save_to_dict` or :meth:`yastn.tn.mps.MpsMpo.save_to_hdf5` can be again deserialized into MPS/MPO.
+MPS/MPO can be saved as Python `dict` or `HDF5` file.
+The MPS/MPO previously serialized by :meth:`yastn.tn.mps.MpsMpo.save_to_dict`
+or :meth:`yastn.tn.mps.MpsMpo.save_to_hdf5` can be again deserialized into MPS/MPO.
 
 Examples of exporting and loading MPS/MPO can be found in :ref:`examples/mps/build:save and load mps/mpo`.
