+++ yastn/tn/mps/__init__.py
@@ -5,6 +5,6 @@ from ._env import Env2, Env3, measure_overlap, measure_mpo, vdot, measure_1site,
 from ._dmrg import dmrg_
 from ._tdvp import tdvp_
 from ._compression import compression_, zipper
-from ._generate import random_dense_mps, random_dense_mpo, random_mps, random_mpo
-from ._generate import Hterm, generate_product_mpo, product_mps, product_mpo
-from ._generate import generate_mpo, generate_mpo_template, generate_mpo_fast, Generator
+from ._initialize import random_dense_mps, random_dense_mpo, random_mps, random_mpo, product_mps, product_mpo
+from ._generate_mpo import Hterm, generate_mpo, generate_mpo_preprocessing, generate_mpo_fast
+from ._generator_class import Generator
